Use __int64 instead of SSIZE_T -- don't know where latter is defined
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@1352 98333e67-4a24-44d7-a75c-e53540dd3050

+++ trunk/lib/abyss/src/file.c
@@ -51,7 +51,7 @@
 
 #if MSVCRT
   #include <io.h>
-  typedef SSIZE_T readwriterc_t;
+  typedef __int64 readwriterc_t;
 #else
   #include <unistd.h>
   #include <fcntl.h>
